Allow Cody to autoupdate itself in JetBrains (#7797)
## Changes

Add autoupdate option for Cody plugin.
User can choose to let Cody update automatically, get the notification,
or skip update check alltogether.


![image](https://github.com/user-attachments/assets/0428861d-d45d-4f92-b496-aaf854ab3115)

## Test plan

1. Open IntelliJ and install new Cody plugin
2. Just after startup and then every after every 4h you should see on
the status bar info that Cody is checking for updates (it is pretty
quick so might not be easy to spot - it takes around 2s).

It is a bit hard to check in the wild because at the time of testing
that Cody plugin will be newest, so update won't be triggered. I will
make sure to test it well enough locally with a locally hacked
versioning.