fix(autoedit): Use correct `position` for hot streak chunks (#7811)
## Description

Previously we would always force the `position` to be at the start of
the hot-streak chunk, and then adjust the prefix to be empty to account
for it. That has some issues:
1. We incorrectly make inline completions for the first chunk, as the
`position` is different than the users actual cursor position. This
means sometimes we _think_ we have shown a completion but it was
actually hidden by VS Code. We should have shown an edit in this case
instead.
2. We need to also adjust other values like `maxPrefixLength` and
`maxSuffixLength` when we reduce the prefix. This can get complex and
it's easier just to maintain the original values (although we may want
this logic in the future.)

closes https://github.com/sourcegraph/cody/pull/7811

## Test plan

Manual testing on real examples and mocked examples in cody-chat-eval

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->