feat(agentic chat): Use Gemini Flash mode as default
Close https://linear.app/sourcegraph/issue/CODY-5736 https://linear.app/sourcegraph/issue/CODY-5781

This commit prioritizes the Gemini Flash > Haiku 3.5 models for the DeepCodyAgent.

If the `useDefaultChatModel` feature flag is disabled, the agent will now prefer to use Gemini Flash by default, or Haiku 3.5 if  Gemini Flash is not available. If neither is available, it will fall back to the first model tagged with "speed". This ensures that the agent uses a fast model when possible, improving the user experience.

## Test Plan

Added unit test for setting the right model for agentic chat.