fix(prompt-editor): Address flushSync warning when setting editor state
This commit addresses the `flushSync` warning that occurs when setting the editor state in the PromptEditor component. The issue arises because the editor state is being updated within React's render cycle.

To resolve this, `requestAnimationFrame` is used to ensure that the editor state is set outside of React's render cycle. This change is applied in two places:

- When setting the editor state from serialized data.
- When initializing the editor state with the initial editor state.