feat(agentic chat): Use Gemini Flash mode as default (#7831)
Close https://linear.app/sourcegraph/issue/CODY-5736
https://linear.app/sourcegraph/issue/CODY-5781

This commit prioritizes the Gemini Flash > Haiku 3.5 models for the
DeepCodyAgent.

If the `useDefaultChatModel` feature flag is disabled, the agent will
now prefer to use Gemini Flash by default, or Haiku 3.5 if Gemini Flash
is not available. If neither is available, it will fall back to the
first model tagged with "speed". This ensures that the agent uses a fast
model when possible, improving the user experience.

## Test Plan

Added unit test for setting the right model for agentic chat.

1. Connect to S2 and ask Cody a question using agentic chat
2. Verify agentic context is used and work

Example: `why isn't the cody commit message generator using the git
template in this codebase?`


![image](https://github.com/user-attachments/assets/f4ee4f35-0be6-4d0d-9266-5e9030d16c38)