feat(chat): add copy message button to copy message output (#7807)
Fixes CODY-4963 introduce Copy message button to allow copying of
message output useful especially on Cody web.


![image](https://github.com/user-attachments/assets/8cd4a1ed-b0ce-49cc-a20a-b3a48aaeb670)

The copy button at the bottom corner of the AssistantMessageCell and it
works both in VSCode and in Cody web.

Most copilots have such a copy button to get the output which is super
useful, especially if you want to pass the reply somewhere else for further processing.

## Test plan
- Verify that the copy button is displayed correctly in the assistant
message cell.
- Verify that clicking the copy button copies the assistant's message to
the clipboard.
- Verify that the copy button is not displayed when the message is
loading, has errors (unless aborted), or is a search intent.