fix(auto-edit): use correct doc context version to show inline completions more often (#7837)
- Renamed `docContext` to `predictionDocContext` and `requestDocContext`
to clearly differentiate between them in the code.
- In `vscode/src/autoedits/renderer/manager.ts`, the `getRenderOutput`
function used to rely on `requestDocContext` to prepend the current line
prefix to the inline completion insert text. However, if a user added
any text after the completion started, and the current line prefix
changed, it led to the `insertText` not matching the text already in the
document and being hidden by VS Code. This PR fixes the issue by
removing the dependency on `docContext` entirely from this function.
Instead, it now relies on the always up-to-date recomputed current line
prefix and suffix.
- Inlines the `getAndProcessModelResponses` body into `getPrediction` to
reduce the number of nested functions for easier debugging.

(cherry picked from commit 771de02b0e3b79a9c9cf6365275471da9680c967)