fix(chat): Hide copy button in non-chat mode (#7895)
This commit hides the copy button in the assistant message cell when not
in chat mode. The copy button is only relevant when the assistant is
responding to a chat-based query. This change ensures that the copy
button is not displayed when the assistant is used for other purposes,
such as edit or search or agent.


## Test plan

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

Use non chat mode and confirm the copy button is not showing up