Properly use generated protocol classes (#7922)
## Changes

This PR switches us to fully use generated protocol and makes sure we
correctly implement all interfaces.
It also allows us to remove few hundred lines of boilerplate.

It removes couple of `CurrentConfigFeatures` classes which were only
used to observe changes to `serverSentModels` flag (needed in
[LlmDropdown](https://github.com/sourcegraph/cody/pull/7922/files#diff-142b743d497e68d7462b4eabbf071688bb3e4c31a719c9bf147b7b89fe51d5be)).
But it was done incorrectly, and this code in fact was not working.

It started to be broken almost half year ago and no one even noticed.
Original cause was refactoring done in [this
PR](https://github.com/sourcegraph/cody/pull/6182/files#diff-96fcfc91b38ef7dec1c22931edfd74464efb663547e49be8718ecea3171113b8)
which changed the way we were passing `clientConfig`, and removed
`serverSentModels`. We should use `modelsAPIEnabled` instead.

I'm going to fix it, but it can be implemented in much simpler way, so
since it was not working for a long time I'm removing all that bloated
classes handling it, and I will soon provide much simpler fix. Not in
this PR though, because I have first to consult current usage and
meaning of `modelsAPIEnabled` (I sent question about it to
@jamesmcnamara).

## Test plan

N/A

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->