Add handling for read-only models list in jetbrains edit window (#7935)
## Changes

This PR brings back functionality to disable models list selection if
account is enterprise and `clientConfig?.modelsAPIEnabled` is `false`.
This was broken several months ago and no one noticed:


https://github.com/sourcegraph/cody/pull/6182/files#diff-96fcfc91b38ef7dec1c22931edfd74464efb663547e49be8718ecea3171113b8

I recently removed our code for handling it (since it was not working
anyway and was overly complicated) and now bringing back simpler version
of it.

## Test plan

Hard to test without being able to change server flags, but I changed
server response in debugger and verified it works.

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->