Bump min version for new prompt api (#7943)
closes:
https://sourcegraph.slack.com/archives/C04MSD3DP5L/p1747342638876129
context:
https://sourcegraph.slack.com/archives/C04MSD3DP5L/p1747458746396879?thread_ts=1747342638.876129&cid=C04MSD3DP5L

We made changes to the GQL API particularly
[this](https://github.com/sourcegraph/sourcegraph/pull/5211), and merged
it before the Sourcegraph 6.3 patch release cut on main. The Cody client
does check for the site version to be > 6.3.1 to use the new query.

But the latest Sourcegraph patch release 6.3.2x, on
[demo.sourcegraph.com](http://demo.sourcegraph.com/) doesn't contain the
API changes we made and merged on main earlier because the changes had
to be backported to the 6.3 release branch to make it part of the
Sourcegraph patch release.

Now as part of this PR we bump up the minimum version required on the
Cody client to 6.4 so that the new query is only used with the next SG
release cut from main.

## Test plan

- connect to demo.sourcegraph.com
- make sure the prompt picker displays all prompts and not just the
inbuilt ones.

(cherry picked from commit 55aca8bf0e4553b3fe2fe56c3802b5abc8905726)