revert(chat): enable agentic chat based on feature flag and model (#7951)
This commit reverts the changes made in
https://github.com/sourcegraph/cody/pull/7761

- Updates the `isAgenticChatEnabled` method in `ToolboxManager` to check
for a feature flag (`FeatureFlag.DeepCody`) and the specific model being
used (`DeepCodyModelRef`). This ensures that agentic chat is only
enabled when the feature flag is enabled and the selected model is the
Deep Cody model.
- Modifies the `syncModels` function to add the Deep Cody model to the
list of available models if the `DeepCody` feature flag is enabled and
the model is not already present.
- Updates `ChatController` and `registry` to pass the model to
`isAgenticChatEnabled` to ensure the agent is only used with the correct
model.
- Removes unnecessary check for `toolboxManager.isAgenticChatEnabled()`
in `DeepCodyHandler` as the check is now done earlier.
- Adds `DEEP_CODY_MODEL` and `DeepCodyModelRef` to
`lib/shared/src/models/client.ts`

These changes ensure that agentic chat is only enabled for users who
have the feature flag enabled and are using the Deep Cody model. This
allows us to control the rollout of agentic chat and ensure that it is
only used by users who are intended to use it.

## Test plan

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

- Verify that agentic chat is only enabled when the `DeepCody` feature
flag is enabled and the Deep Cody model is selected.
- Verify that agentic chat is disabled when the `DeepCody` feature flag
is disabled or a different model is selected.
- Verify that the Deep Cody model is added to the list of available
models when the `DeepCody` feature flag is enabled and the model is not
already present.

Verify agentic chat shows up in model dropdown:

![image](https://github.com/user-attachments/assets/1655f1ad-84a4-4f24-954c-82d3fe4ad928)

Ask agentic chat: "How do we display speaker transcript in the UI in
this codebase"

![image](https://github.com/user-attachments/assets/1cfdd366-b50a-4fa4-9bf7-a2798fb5528c)

Ask the same question to Claude 3.7 Sonent and confirmed no agentic
context was added:

![image](https://github.com/user-attachments/assets/9944154a-65e6-4492-afae-f0fe8e5ebead)

(cherry picked from commit 8a08165895adc5251d7e8d2f84bca8a41ca46b8e)