Add config.condebase check to repoNameResolver (#7944)
Issues: pending link 

Customers have proxy repo origins configured on git but they do not have
a way to set a custom correct/original/real remote for the codebase so
that Cody gets context from Sourcegraph using the right repo name and
not the proxy one.

Since beginning we have `config.codebase` for this purpose but that has
not been integrated properly since the inception of `repoNameResolver`.

We were checking the `config.codebase` only inside the
`gitMetadataForCurrentEditor` which is only used in inline completions
and nowhere else. In the entire codebase we use repoNameResolver
directly.

This PR moves the `config.codebase` check to the `repoNameResolver` so
that it is integrated with the entire codebase.

## Test plan

Follow the demo:
https://www.loom.com/share/7cfb6fca7c1944158b9fadb1cc460664