Do not hide stable releases by default (#7813)
Some time ago we decided to hide stable releases by default. That was
caused by the fact we were focused on limiting the time from the moment
releases to the moment of the JB approval (having QA in between).
Presently, we are more careful about stable releases. We can be more
confident about stables. We can make it public by default.

## Test plan
- n/a

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->