Do not trigger autocompletion on Delete (#7990)
See the related discussion:
https://sourcegraph.slack.com/archives/C04MSD3DP5L/p1747722947410229.

Personally, I use Delete almost never 😅 it turned out that in case of
having an autocompletion displayed and triggering delete it was removing
the autocompletion and retriggering it again. Let's simply skip the
trigger in that case.

## Test plan
1. Having an autocompletion displayed 
2. Hit Delete (fn + Backspace) -> Removes autocomplete suggestion
3. Hit Delete (fn + Backspace) -> Removes next character which is a
newline in our case

Also, approved by the reported.

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->