Fix config management in agent and jetbrains (#7989)
## Changes

This PR fixes bugs in config update code. There were two independent
bugs, one in the agent and one in jetbrains.
After this PR it is possible to add MCP in JetBrains using MCP
management UI and changes are reflected in the config:


![image](https://github.com/user-attachments/assets/2cb199a9-4b25-452c-a10e-27259a52d0cf)


## Test plan

Automatic tests added.

Other than that please test adding MCP server in JetBrains as in MCP QA
guide provided by @abeatrix
(I'm not sharing it there as it is not public)

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->