fix(jetbrains): remove unset SKIP_CODE_SEARCH_BUILD env var
Removes the code to unset the `SKIP_CODE_SEARCH_BUILD` environment variable from the `verify-release.sh` script.  This code is causing the release script to fail locally even when `SKIP_CODE_SEARCH_BUILD`  sets to true. The GITHUB_TOKEN should be set in the CI only and should not required release captain to set it locally for security reason