remove duplicated parsed text as content (#8011)
we previously stream the completion to the callback function as text and
also as text content which is redundant and would cause confusion to the
LLM when both are sent back as part of chat history.

also revert the sanitize refactoring that is causing a bug in chat.

## Test plan

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

all tests are green. try agent mode to confirm no text is being repeated