Split edit code between flow/model and UI (QuickPick) (#8013)
Initial work for
https://linear.app/sourcegraph/issue/QA-668/core-method-not-implemented-error-display-when-using-edit-search

## Changes

Intention of this PR is mostly to prepare codebase for sub-sequent PRs
which will unify the way in which VSC and other clients use the edit
code. Currently `getInput` method which is at the core of UI logic is
not called by clients at all, and instead edit code is hacked a bit in
the agent. In the long run that can lead to various issues like this
one:
https://linear.app/sourcegraph/issue/QA-668/core-method-not-implemented-error-display-when-using-edit-search

One side effect of this PR is that I fixed title of the quickpic, which
is updated if user changes selection type (e.g. from range to line).

## Test plan

Requires full edits QA.

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->