fix(chat): chat transcript scrolling (#8020)
Improvements to the chat transcript scrolling behavior for the chat
webviews.

- Implements logic to prevent auto-scrolling when the user has manually
scrolled away from the bottom of the transcript. This ensures that the
user's position is maintained when they are reviewing previous messages.
- The auto-scroll behavior is reset when a new message is finished
streaming.
- Hides the scrollbar in the chat transcript.

These changes provides more control over the chat transcript and
preventing disruptive auto-scrolling which was the previous behavior.


## Test plan

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

Ask Cody a question and verifyed:
- autoscroll as message is streaming
- you can break autoscroll if you scroll while the message is streaming
- scroll to bottom shows up if you're not at the bottom