Restoration of unit test generation, path normalization (#7765)
This PR adds a class that attempts to normalize all the Uri with file
paths that we get from the agent. Non-standard formatted uri causes
inability to create file in editor (ex. file with unit tests). In
addition, the editor does not support the creation of new files in
memory by which the use of the prefix untitled in uri was incorrectly
interpreted as a request to create a scratchpad file. This PR causes
that if the agent request that a file needs to be created, it will be
created as a project file (not a scratch file).

## Test plan
**Scenario I**
1. Open file with code in JB editor
2. Select code and open context menu
3. Run Generate unit test command
4. New test file should appear

**Scenario II**
1. On Windows start JB editor
2. From sidebar select Remote Development -> WSL
3. Use + button to start wizard to configure new Ubuntu environment 
5. Select project to edit
6. Start editing remote project
7. Select code and open context menu
8. Run Generate unit test command
9. New test file should appear
<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

---------

Co-authored-by: Tomasz Gołębiowski <tgolebiowski@virtuslab.com>