Unify inline edit architecture in VSC and other clients (#8023)
Fixes
https://linear.app/sourcegraph/issue/QA-668/core-method-not-implemented-error-display-when-using-edit-search

## Changes

This PR changes inline edit architecture in JetBrains.

Previously we were doing a bit of the state management on the JetBrains
side. It was especially visible for actions like retry, where we had to
get task details to obtain the prompt text.
Implementation of inline edits was a bit different then in VSC. For
example edit range management in JetBrains was always explicit (edit was
happening for current selection or current line) while in VSC if there
was no selection, most relevant scope was used. Now JetBrains reuses the
same logic through agent, and I added endpoints which allow agent to
select and reveal text ranges. This part will also help in some other
places where we are already using that functionality in VSC
(`revealRange` is used in a dozen of places).

I also simplified agent API to pass less data, and removed some
deprecated endpoints which were leaking internal agent state
(`editTask/didUpdate`, `editTask/didDelete`).

Despite adding some features to existing functionality this PR manages
to shave off ~100 lines of the code.

## Test plan

Full QA for inline edit functionality is required.

Testplan for the bugfix:

Steps to Reproduce:

1. Log in to the Cody plugin using any available account (Free, Pro, or
Enterprise).
2. From the search intent dropdown menu, select "Edit" option.
3. Press Enter to execute the search.
4. Edit prompt window should appear.
5. Fill the edit request and confirm.
6. Edit should be applied.

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->