chore(test): Skip flaky auth test on Ubuntu (#7997)
CONTEXT:
https://sourcegraph.slack.com/archives/C05AGQYD528/p1748038226893879

The `switches to a different account` test in `auth.test.ts` was flaky
on Ubuntu 18 and has been failing for a few months on main. This commit
skips the test on that platform to prevent spurious test failures.


## Test plan

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

Green CI. The test should still run in non-ubuntu unit test CI