feat(chat): improve scrollbar marker bottom scroll detection
This commit refactors the scrollbar marker logic to use a tolerance value when determining if the user has scrolled to the bottom of the container. This prevents the markers from updating unnecessarily when the user is very close to the bottom of the scrollable area, especially after using "Skip to end".

The following changes were made:

- Added a `BOTTOM_SCROLL_TOLERANCE_PX` constant to `MARKER_CONFIG` with a value of 5px.
- Created `isScrolledToBottom` helper function to encapsulate the bottom scroll check logic, using `BOTTOM_SCROLL_TOLERANCE_PX`.
- Replaced the hardcoded `5` pixel tolerance with the `isScrolledToBottom` helper function call in the scroll and mutation observer callbacks.
- Added `isScrolledToBottom` to the dependency array of `useCallback`.