Fix auto-edit's caching issues (#8045)
I noticed a problem with auto-edits when triggering manually. This is a
fix. My changes make sure that we update the last trigger timestamp for
auto-edits so it is later properly used to determine whether it is a
manual or invoke trigger type, and then whether to use the caching
mechanism.

## Test plan
- trigger auto-edit manually 

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->