New capability to hide all action buttons except 'Copy Code' (#8072)
This PR adds the ability to hide action buttons under each code block in
the chat window. This is temporarily needed for Visual Studio which does
not support actions such as 'Smart Apply' or 'Save to file'. Once this
capability is enabled, only the 'Copy Code' to clipboard button will be
available. Once 'Smart Apply' will be implemented in VS, these changes
can be reverted.
## Test plan
1. Enable hiding using `codeCopyOnlyAction` capability 
2. Check that only 'Copy code' option is available under each code block
<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

---------

Co-authored-by: Tomasz Gołębiowski <tgolebiowski@virtuslab.com>