Do not hide lookups (#8073)
This PR is a minor follow up to
https://github.com/sourcegraph/cody/pull/8071.
Accidentally I made lookups hint disappear just after they are shown.
That is clearly wrong. This PR makes sure that we dispose hints only if
there is an active autoedit.

Also, the only reasonable way to dismiss the auto edit popup is to use
the message bus - this PR does this change.

## Test plan
- lookups work
- lookup hint does not show with auto edit popup


<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->