Fix JetBrains uri parsing issues (#8065)
Fixes
https://linear.app/sourcegraph/issue/QA-700/jb-onlywindows-only-cody-gets-crashed-when-user-tries-to-save-searched

## Changes

### Bugfix

Before my changes the test I added was failing with:

```
Caused by: java.net.URISyntaxException: Illegal character in path at index 22: 
file:///C:/Users/Rajat Nagoria/PycharmProjects/pythonProject2/Untitled\

Illegal character in path at index 24: file:///C:/dev/Jetbrains Tests Projects/kotlin/TestProject/Validate.kt
java.net.URISyntaxException: Illegal character in path at index 24: file:///C:/dev/Jetbrains Tests Projects/kotlin/TestProject/Validate.kt
	at java.base/java.net.URI$Parser.fail(URI.java:2976)
	at java.base/java.net.URI$Parser.checkChars(URI.java:3147)
	at java.base/java.net.URI$Parser.parseHierarchical(URI.java:3229)
```

It was cause by usage of `URLDecoder.decode` which e.g. converts `%20`
to spaces, which are illegal in the URI.
We should not use it at all in this specific place. I fixed that and
added few other tweaks.

### Architecture changes

In the process of fixing the bug I came to the conclusion that whole
`untitled` documents handling we had in the agent was error prone, not
fully correctly implemented, and in fact not really needed. If we make
sure we only show saved files to the user (and we have control over
that) we can completely remove quite a bit of custom logic. I removed
`textDocument/openUntitledDocument` endpoint, and instead implemented
`AgentTextDocument::save` and used it in one additional place.

That decision reduces a lot of the complexity on the clients/agent side,
but also have one downside.
There is no easy way to ensure at compile time that we never try to show
an untitled document. We could add a logic for checking that in the
`showTextDocument` but we cannot handle it in any reasonable way. We
could try to save file there first if it is untitled, but untitled file
is not guaranteed to have reasonable name/path.
Ultimately it might be just easier and better to QA our edit logic to
make sure it does not try to show such files, and do not overcomplicated
the implementation otherwise.

## Test plan

Unit test covering problematic scenario was added.

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->