Do not copy code to clipboard on the EDT thread (#8085)
## Changes

I verified that adding a sleep next to
`CopyPasteManager.getInstance().setContents` call causes IDE to hang.
We do not control if that call is blocking or not, but we can run it in
a background thread.

That will prevent hangs, but in case of slow `setContents` call it may
cause delay in clipboard copying operation (text in the clipboard may
not be available immediately).
But it is better alternative to hang, and in the majority of cases that
operation is instant.

## Test plan

1. Generate some code with Cody chat
2. Click "Copy" button under the snapshot
3. Do it a few times
4. No hands should be observed


<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->